+++ src/OpenTelemetry/CHANGELOG.md
@@ -10,12 +10,25 @@
 * `Resource` is no longer added to observed `Activity` objects as a
   `CustomProperty`.
   ([#1463](https://github.com/open-telemetry/opentelemetry-dotnet/pull/1463))
-* Remove RentrantExportProcessor as it is not required by spec.
+* Removed `RentrantExportProcessor` as it is not required by spec.
 * `ActivitySourceAdapter` supports setting `ActivitySource` for Activities
   created without `ActivitySource`.
   ([#1515](https://github.com/open-telemetry/opentelemetry-dotnet/pull/1515/))
-  * Implemented Shutdown for TracerProvider
-    ([#1489](https://github.com/open-telemetry/opentelemetry-dotnet/pull/1489))
+* Implemented `Shutdown` for `TracerProvider`.
+  ([#1489](https://github.com/open-telemetry/opentelemetry-dotnet/pull/1489))
+* `Resources.CreateServiceResource` has been removed in
+  favor of the `ResourceBuilder` API.
+  ([#1533](https://github.com/open-telemetry/opentelemetry-dotnet/pull/1533))
+* `TracerProviderBuilder.SetResource` has been changed to
+  `TracerProviderBuilder.SetResourceBuilder`.
+  ([#1533](https://github.com/open-telemetry/opentelemetry-dotnet/pull/1533))
+* By default `TracerProvider` will set a `Resource` containing [Telemetry
+    SDK](https://github.com/open-telemetry/opentelemetry-specification/tree/master/specification/resource/semantic_conventions#telemetry-sdk)
+    details
+    ([#1533](https://github.com/open-telemetry/opentelemetry-dotnet/pull/1533)):
+  * `telemetry.sdk.name` = `opentelemetry`
+  * `telemetry.sdk.language` = `dotnet`
+  * `telemetry.sdk.version` = [SDK version]
 
 ## 0.8.0-beta.1
 
@@ -24,7 +37,6 @@ Released 2020-Nov-5
 * TracerProviderBuilder API changes Renamed AddInstrumentation to
   AddDiagnosticSourceInstrumentation and made internal. Added AddInstrumentation
   ([#1454](https://github.com/open-telemetry/opentelemetry-dotnet/pull/1454))
-
 * DiagnosticSource subscription helper classes (DiagnosticSourceSubscriber,
   ListenerHandler,PropertyFetcher) are made internal.